do some refactor
