--- src/main/java/seedu/duke/deliveryOrder.java
@@ -1,51 +1,0 @@
-import java.util.List;
-
-public class deliveryOrder extends Order {
-
-    private final static double carrierFee = 0.3;
-    private final static double deliveryFee = 2.0;
-
-    private boolean isDeliveryOrder;
-    private String address;
-    private String phoneNumber;
-
-    deliveryOrder(Canteen canteen, Stall stall, List<Dish> dishes, Customer customer, boolean isDeliveryOrder,
-                  String address, String phoneNumber) {
-        super(canteen, stall, dishes, customer);
-        this.isDeliveryOrder = isDeliveryOrder;
-        this.address = address;
-        this.phoneNumber = phoneNumber;
-    }
-
-    public boolean getIsDeliveryOrder() {
-        return this.isDeliveryOrder;
-    }
-
-    public double getCarrierFee() {
-        return this.carrierFee;
-    }
-
-    public double getDeliveryFee() {
-        return this.deliveryFee;
-    }
-
-    public String getAddress() {
-        return this.address;
-    }
-
-    public String getPhoneNumber() {
-        return this.phoneNumber;
-    }
-
-    @Override
-    public String toString() {
-        if (isDeliveryOrder) {
-            return super.toString() + "(Delivery Order: address is " + this.address + " phone number is "
-                    + this.phoneNumber + ")";
-        }
-        else {
-            return super.toString() + "(Not Delivery Order)";
-        }
-    }
-
-}
