+++ src/main/java/takeAwayOrder.java
@@ -1,12 +1,13 @@
+import java.util.List;
+
 public class takeAwayOrder extends Order {
 
     private boolean isTakeAway;
-    private double carrierFee;
+    private final static double carrierFee = 0.3;
 
-    takeAwayOrder(Canteen canteen, Stall stall, Dish dish, boolean isTakeAway, double carrierFee) {
-        super(canteen, stall, dish);
+    takeAwayOrder(Canteen canteen, Stall stall, List<Dish> dishes, Customer customer, boolean isTakeAway) {
+        super(canteen, stall, dishes, customer);
         this.isTakeAway = isTakeAway;
-        this.carrierFee = carrierFee;
     }
 
     public boolean getIsTakeAway() {
@@ -15,5 +16,15 @@ public class takeAwayOrder extends Order {
 
     public double getCarrierFee() {
         return this.carrierFee;
+    }
+
+    @Override
+    public String toString() {
+        if(isTakeAway) {
+            return super.toString() + "(Take Away)";
+        }
+        else {
+            return super.toString() + "(Not Take Away)";
+        }
     }
 }
